guessnumber.py
The aim of the game is for the player to guess the number that has been randomly generated. The final score of the player will be determined by the number of attempts it took the individual to get to the correct answer. The lesser the number of tries, the higher the value of the score. Upon each wrong guess, a hint will be provided to the player, stating whether they are above the number, below the number, or whether the number is a multiple of the wrongly guessed answer.