print a collection
Below you see three objects which are of collection data type: a list, a tuple, 
and a string. Although different in many ways, you can write a 'for' loop that 
steps through a "collection" and it will work with all three. This problem
is started for you. Finish it by writing the loop that will step through the
collection and print out its items, one per line.